:minibus: Move rule validation helpers to rules.js
